Updated the player models to be document based instead of relational databases

+++ services/server/endpoints/player.py
@@ -1,13 +1,13 @@
 from flask import request
 from werkzeug.exceptions import BadRequest, Unauthorized, NotFound
 
-from lib.model.models import PlayerEquipmentModel, SpellModel, PlayerProficiencyModel, PlayerModel
-from lib.service import player_service, campaign_service
-from lib.user_session import session_user, session_user_set
 from endpoints import api, json_api, require_login
+from lib.model.models import PlayerModel
+from lib.service import player_service, campaign_service
+from lib.user_session import session_user
 
 
-def check_player(player: PlayerModel):
+def check_player_ownership(player: PlayerModel):
     """
     Checks whether or not a player exists, and whether or not it belongs to the logged in user.
 
@@ -32,56 +32,26 @@ def delete_player(player_id):
         :param player_id: The id of the player that will be deleted.
         :return:
     """
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     player_service.delete_player(player)
-
-    return {
-        "success": True,
-    }
+    return player.to_json()
 
 
-@api.route('/player/<int:player_id>/campaign', methods=["PUT"])
-@json_api()
-@require_login()
-def set_player_campaign(player_id):
-    player = player_service.find_player(player_id)
-    check_player(player)
-
-    data = request.get_json()
-
-    required_fields = ["campaign_code"]
-
-    if not data or (False in [x in data for x in required_fields]):
-        raise BadRequest()
-
-    campaign = campaign_service.find_campaign_with_code(data.get("campaign_code"))
-    if campaign is None:
-        raise NotFound("This campaign does not exist.")
-
-    player_service.update_player_campaign(player, campaign.id)
-
-    return {
-        "success": True
-    }
-
-
-@api.route('/player/<int:player_id>/data', methods=["GET"])
+@api.route('/player/<int:player_id>', methods=["GET"])
 @json_api()
 @require_login()
 def get_player(player_id):
-    player = player_service.find_player(player_id)
-    check_player(player)
+    """
+        Fetches a player with the given player_id in the URL.
 
-    player_info = player_service.get_player_info(player)
-    player_proficiencies = player_service.get_player_proficiencies(player)
-
-    data = player.to_json()
-    data["info"] = player_info.to_json()
-    data["proficiencies"] = player_proficiencies.to_json()
-
-    return data
+        :param player_id: The id of the player that will be deleted.
+        :return:
+    """
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
+    return player.to_json()
 
 
 @api.route('/player/<int:player_id>', methods=["PUT"])
@@ -101,35 +71,38 @@ def set_player_info(player_id):
          - backstory: The new backstory for your PC
          - info: Json containing player info
          - proficiencies: Json containing player proficiencies
-        :return: A json object containing the updated player's id.
+        :return: The updated player model
     """
     data = request.get_json()
 
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
-    # Update main player information
-    player_service.update_player(player, data.get("name"), data.get("race"), data.get("class_ids"),
-                                 data.get("backstory"), data.get("money"))
+    return player_service.update_player(player, data)
 
-    info = data.get("info")
-    if info is not None:
-        player_service.set_player_info(player, info.get("strength"), info.get("dexterity"), info.get("constitution"),
-                                       info.get("intelligence"), info.get("wisdom"), info.get("charisma"),
-                                       info.get("saving_throws_str"), info.get("saving_throws_dex"),
-                                       info.get("saving_throws_con"), info.get("saving_throws_int"),
-                                       info.get("saving_throws_wis"), info.get("saving_throws_cha"), info.get("max_hp"),
-                                       info.get("armor_class"), info.get("speed"), info.get("level"))
 
-    profs = data.get("proficiencies")
-    if profs is not None:
-        player_service.update_proficiencies(player, profs)
+@api.route('/player/<int:player_id>/campaign', methods=["PUT"])
+@json_api()
+@require_login()
+def set_player_campaign(player_id):
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
-    # TODO: Maybe return something more useful here.
+    data = request.get_json()
+
+    required_fields = ["campaign_code"]
+
+    if not data or (False in [x in data for x in required_fields]):
+        raise BadRequest()
+
+    campaign = campaign_service.find_campaign_with_code(data.get("campaign_code"))
+    if campaign is None:
+        raise NotFound("This campaign does not exist.")
+
+    player_service.update_player_campaign(player, campaign.id)
+
     return {
-        "success": True,
-        "updated_info": info is not None,
-        "updated_proficiencies": profs is not None
+        "success": True
     }
 
 
@@ -143,8 +116,8 @@ def add_player_item(player_id):
     if item_id is None:
         raise BadRequest("No item_id specified.")
 
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     player_item = player_service.player_set_item(player, item_id, data.get("amount", 1))
 
@@ -157,8 +130,8 @@ def add_player_item(player_id):
 def update_player_item(player_id, item_id):
     data = request.get_json()
 
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     player_item = player_service.player_set_item(player, item_id, data.get("amount"), data.get("extra_info"))
 
@@ -172,8 +145,8 @@ def add_player_spell(player_id):
     data = request.get_json()
     spell_id = data.get("spell_id", None)
 
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     player_spell = player_service.player_add_spell(player, spell_id)
     return player_spell.spell.to_json()
@@ -183,8 +156,8 @@ def add_player_spell(player_id):
 @json_api()
 @require_login()
 def get_player_items(player_id):
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     player_items = player_service.get_player_items(player)
 
@@ -197,8 +170,8 @@ def get_player_items(player_id):
 def delete_player_spell(player_id, spell_id):
     user = session_user()
 
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     player_spells = player_service.delete_player_spell(user, player, spell_id)
     return [player_spell.spell.to_json() for player_spell in player_spells]
@@ -210,8 +183,8 @@ def delete_player_spell(player_id, spell_id):
 def delete_player_item(player_id, item_id):
     user = session_user()
 
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     error = player_service.delete_player_item(user, player, item_id)
 
@@ -225,8 +198,8 @@ def delete_player_item(player_id, item_id):
 @json_api()
 @require_login()
 def get_player_spells(player_id):
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     player_spells = player_service.get_player_spells(player)
 
@@ -245,15 +218,9 @@ def get_available_spells():
 @json_api()
 @require_login()
 def get_player_proficiencies(player_id: int):
-    player = player_service.find_player(player_id)
-    check_player(player)
-
-    player_proficiencies: PlayerProficiencyModel = player_service.get_player_proficiencies(player)
-
-    return {
-        "success": True,
-        "proficiencies": player_proficiencies.to_json()
-    }
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
+    return player.info.get("proficiencies")
 
 
 @api.route('/player/<int:player_id>/proficiencies', methods=["PUT"])
@@ -261,39 +228,28 @@ def get_player_proficiencies(player_id: int):
 @require_login()
 def set_player_proficiencies(player_id: int):
     data = request.get_json()
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
-    player = player_service.find_player(player_id)
-    check_player(player)
-
-    player_service.update_proficiencies(player, data)
-
-    return {
-        "success": True
+    # Ensure it's wrapped correctly
+    proficiencies = {
+        "proficiencies": data
     }
+    # TODO: Fix
+    player_service.update_player(player, data)
+    return player.info.get("proficiencies")
 
 
 @api.route('/player/<int:player_id>/classes', methods=["GET"])
 @json_api()
 @require_login()
 def get_player_class(player_id):
-    player = player_service.find_player(player_id)
-    check_player(player)
+    player = player_service.get_player(player_id)
+    check_player_ownership(player)
 
     class_models = player_service.get_classes(player)
-    classes = []
-    for class_model in class_models:
-        abilities = player_service.get_class_abilities(class_model)
-        classes.append({
-            "id": class_model.id,
-            "name": class_model.name,
-            "info": class_model.info,
-            "table": class_model.table,
-            "abilities": [ability.to_json() for ability in abilities]
-        })
-    return {
-        "success": True,
-        "classes": classes
-    }
+
+    return [cls.to_json() for cls in class_models]
 
 
 print("Loaded player endpoints")
